Remove messages from IO testing
IO testing was performed using the messages as elements of a vector and
then counting them. This is not trustable when using multithreading or
big amounts of data because messages that are expected to be independent
by the tests can be sent into the same buffer.

Now the tests are based on the total bytes read. Making it more
trustable and bombproof.